requirements.txt
Flask==2.2.3
flask_cors==3.0.10
docker_py==1.10.6
pymongo==4.3.3
python-dotenv==1.0.0
redis==4.5.2
requests==2.28.0